# QUIC: Fix send buffer sizing to avoid pathological performance under high RTT