@kroeckx 2023-10-31
I have to wonder if we need to make this configurable. 6 MB per connection seems high and you can probably allocate all RAM with a few 1000 connections.