@hlandau 2023-10-31
We should make it configurable in the future (compare Linux sysctls `net.core.[rw]mem_max`, `net.ipv4.tcp_[rw]?mem`).

Worth noting that Linux uses default maximums of 4 MiB on the write side, 6 MiB on the read side.